Remove superfluous pointer indirection from prepared statement
Holding the native prepared statement pointer in a separate handler
which was itself wrapped in a shared_ptr was superfluous.

In the mysql connector, unraveling this also required moving result data
from the handler into bind_result (which in hindsight is the obvious
thing to do anyways, of course).

While resolving this, also establish const-correctness in prepared_*.h