Make core/name/create_name_tag.h a stand-alone header, #30
Macros are not exported by modules. They need to be included via an
extra header. This header should have minimal dependencies.

This commit removes all dependencies of create_name_tag.h.

This required the removal of sqlpp::name_tag_base. A result of this
change is that instead of calling `cte(sqlpp::alias::a)`, you now have
to call `sqlpp::cte(sqlpp::alias::a)`.