Provide order_by, limit, and offset for union, #83
union args must not contain order_by, limit, or offset[*].

But these clauses can be appended to a union statement.
Note that column names must not be table-qualified in the
union_order_by.

[*] Postgresql and MySQL allow these clauses in union args if the union
args and enclosed in parentheses. But sqlite3 does not allow any of
that. So let's go with the most restrictive version.
If you really want to use order_by in a union arg, use sub-queries.